Convert AlKharidGeneralStore plugin from kts to kt
